+++ src/yafowil/utils.py
@@ -3,7 +3,10 @@ import json
 import logging
 import re
 from pkg_resources import iter_entry_points
-from node.utils import UNSET
+from node.utils import (
+    Unset,  # B/C
+    UNSET,
+)
 
 
 def get_entry_points(ns=None):